requirements.txt
Flask==2.0.2
numpy==1.21.5
Pillow==9.0.1
pygame==2.1.2
